Some fix on initializing and updating the diffusion reaction rates.

+++ src/Chemistry/Reaction.h
@@ -41,7 +41,7 @@ template <unsigned short M, unsigned short N>
         /// (starting from reactants)
         /// @param rate - the rate constant for this ReactionBase
         Reaction(initializer_list<Species*> species,
-                float rate = 0.0, bool isProtoCompartment = false, float volumeFrac=1.0, int rateVolumeDepPow=0)
+                float rate = 0.0, bool isProtoCompartment = false, double volumeFrac=1.0, int rateVolumeDepPow=0)
             : ReactionBase(rate, isProtoCompartment, volumeFrac, rateVolumeDepPow) {
             initializeSpecies(species);
         }
@@ -266,7 +266,7 @@ private:
 public:
     /// The main constructor
     DiffusionReaction(initializer_list<Species*> species,
-                      float rate = 0.0, bool isProtoCompartment = false, float volumeFrac=1.0)
+                      float rate = 0.0, bool isProtoCompartment = false, double volumeFrac=1.0)
     : Reaction(species, rate, isProtoCompartment, volumeFrac, -1) { // Inversely dependent on compartment active volume
     
         //set averaging
